requirements.txt
# Core LangGraph + LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0  # used with OpenRouter's OpenAI-compatible API
langchain-core>=0.3.0

# PageIndex (vectorless / tree-based PDF retrieval cloud API)
pageindex>=0.2.8

# Local BM25 fallback (used when PAGEINDEX_API_KEY is not set)
rank-bm25>=0.2.2

# PDF text extraction
pypdf>=4.0.0
pdfplumber>=0.11.0

# Streamlit UI
streamlit>=1.38.0

# FastAPI + Uvicorn (for the API server)
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Environment variable management
python-dotenv>=1.0.0

# Utilities
httpx>=0.27.0
